requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
twilio==9.3.0
openai==1.82.0
pydantic==2.10.5
python-multipart==0.0.20
pyjwt==2.9.0
httpx==0.28.1
